Port RecyclerView module to AndroidX and Kotlin
