+++ .travis.yml
@@ -7,9 +7,6 @@ before_install:
   # Install SDK license so Android Gradle plugin can install deps.
   - mkdir "$ANDROID_HOME/licenses" || true
   - echo "d56f5187479451eabf01fb78af6dfcb131a6481e" > "$ANDROID_HOME/licenses/android-sdk-license"
-  # Kotlin binding staleness check.
-  - ./gradlew generateKotlin
-  - if [ ! -z "$(git status --porcelain)" ]; then echo "Bindings changed. Did you run 'generateKotlin'?"; exit 1; fi
   # Install the rest of tools (e.g., avdmanager)
   - sdkmanager tools
   # Install the system image
@@ -18,7 +15,7 @@ before_install:
   - echo no | avdmanager create avd --force -n test -k "system-images;android-18;default;armeabi-v7a"
   - $ANDROID_HOME/emulator/emulator -avd test -no-audio -no-window &
 
-install: ./gradlew ":$MODULE:build" ":$MODULE-kotlin:build" ":$MODULE:assembleAndroidTest" --stacktrace
+install: ./gradlew ":$MODULE:build" ":$MODULE:assembleAndroidTest" --stacktrace
 
 before_script:
   - android-wait-for-emulator
@@ -36,7 +33,7 @@ env:
     - MODULE=rxbinding-appcompat-v7
     - MODULE=rxbinding-design
     - MODULE=rxbinding-leanback-v17
-    - MODULE=rxbinding-recyclerview-v7
+    - MODULE=rxbinding-recyclerview
     - MODULE=rxbinding-support-v4
   global:
     - secure: "kgsT0j2b0GYbQjrDSdIYoh9lXOb2bAqA8gOx7ipegYLiNUuAjTeclw/3IFd5WpPNEhwTvDlb8Pwi/GYsXwYLdjysZUsj73a027swKAuXRBEjnBpfcvhEwKdoO5Ygoz/uH7TvcuCKzEz/bRKIglvn9PcTrzIHHZZIXa/mLU5otNE="
