+++ gradle.properties
@@ -1,5 +1,5 @@
-GROUP=com.jakewharton.rxbinding2
-VERSION_NAME=2.2.1-SNAPSHOT
+GROUP=com.jakewharton.rxbinding3
+VERSION_NAME=3.0.0-SNAPSHOT
 
 POM_DESCRIPTION=RxJava binding APIs for Android's UI widgets.
