--- rxbinding-recyclerview-v7-kotlin/build.gradle
@@ -1,51 +1,0 @@
-buildscript {
-  dependencies {
-    classpath rootProject.ext.androidPlugin
-    classpath rootProject.ext.kotlinPlugin
-  }
-}
-
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-
-dependencies {
-  api project(':rxbinding-recyclerview-v7')
-  api project(':rxbinding-kotlin')
-  api rootProject.ext.kotlinStdlib
-  implementation rootProject.ext.supportAnnotations
-}
-
-android {
-  compileSdkVersion rootProject.ext.compileSdkVersion
-
-  defaultConfig {
-    minSdkVersion rootProject.ext.minSdkVersion
-  }
-
-  compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_8
-    targetCompatibility JavaVersion.VERSION_1_8
-  }
-
-  lintOptions {
-    textReport true
-    textOutput 'stdout'
-  }
-
-  sourceSets {
-    main.java.srcDirs += 'src/main/kotlin'
-  }
-
-  // TODO replace with https://issuetracker.google.com/issues/72050365 once released.
-  libraryVariants.all {
-    it.generateBuildConfig.enabled = false
-  }
-}
-
-apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
-
-tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
-  kotlinOptions {
-    freeCompilerArgs = ['-Xno-param-assertions']
-  }
-}
