--- rxbinding-recyclerview-v7-kotlin/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=rxbinding-recyclerview-v7-kotlin
-POM_NAME=RxBinding Kotlin (recyclerview-v7)
-POM_PACKAGING=aar
