--- rxbinding-recyclerview-v7/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.jakewharton.rxbinding2.support.v7.recyclerview"/>
