+++ rxbinding-recyclerview/gradle.properties
