+++ rxbinding-recyclerview/src/androidTest/AndroidManifest.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<manifest package="com.jakewharton.rxbinding3.recyclerview"
+          xmlns:android="http://schemas.android.com/apk/res/android">
+  <application>
+    <activity android:name="com.jakewharton.rxbinding3.recyclerview.RxRecyclerViewTestActivity"/>
+  </application>
+</manifest>
