+++ rxbinding-recyclerview/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.jakewharton.rxbinding3.recyclerview"/>
