+++ settings.gradle
@@ -12,8 +12,7 @@ include ':rxbinding-design-kotlin'
 include ':rxbinding-support-v4'
 include ':rxbinding-support-v4-kotlin'
 
-include ':rxbinding-recyclerview-v7'
-include ':rxbinding-recyclerview-v7-kotlin'
+include ':rxbinding-recyclerview'
 
 include ':rxbinding-leanback-v17'
 include ':rxbinding-leanback-v17-kotlin'
